+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.72.0]
 
 * Add support for the "Expand selection" command
+* Add "Add predicate/non-terminal declaration" refactoring support
 * Improve documentation of optional settings
 * Fix missing configuration update listener in the `Utils` class